% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asli_app.R
\name{asli_app}
\alias{asli_app}
\title{Run the Shiny Application}
\usage{
asli_app(ui = app_ui, server = app_server)
}
\arguments{
\item{ui}{shiny app_ui() in app_ui.R}

\item{server}{shiny app_server() in app_server.R}
}
\description{
Run the Shiny Application
}
