% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_store.R
\name{get_asli_from_s3}
\alias{get_asli_from_s3}
\title{Get asli output from S3 Bucket}
\usage{
get_asli_from_s3(access_key_id, secret_access_key, endpoint, bucket, key)
}
\arguments{
\item{access_key_id}{S3 access id}

\item{secret_access_key}{S3 secret access key}

\item{endpoint}{S3 host endpoint}

\item{bucket}{S3 bucket name}

\item{key}{filename to target in bucket}
}
\description{
Get asli output from S3 Bucket
}
