% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_asli_plot.R
\name{get_asli_plot}
\alias{get_asli_plot}
\title{Get asli plot}
\usage{
get_asli_plot(s3, year)
}
\arguments{
\item{s3}{S3 Body inherited from object_store()}

\item{year}{the year for which to fetch the image}
}
\description{
Fetch the asli plot png from the s3 bucket for the corresponding year
}
