% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_s3_body.R
\name{get_s3_body}
\alias{get_s3_body}
\title{Get asli output from S3 Bucket}
\usage{
get_s3_body(access_key_id, secret_access_key, endpoint, bucket)
}
\arguments{
\item{access_key_id}{S3 access id}

\item{secret_access_key}{S3 secret access key}

\item{endpoint}{S3 host endpoint}

\item{bucket}{S3 bucket name}
}
\description{
Get asli output from S3 Bucket
}
