% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_asli_year.R
\name{plot_asli_year}
\alias{plot_asli_year}
\title{Plot asli year}
\usage{
plot_asli_year(
  data_dir = "s3://asli",
  mask_filename = "zarr-lsm",
  msl_pattern = "zarr-msl",
  file_name = "asli_calculation_2024.csv",
  year = 2024
)
}
\arguments{
\item{data_dir}{Path to data directory}

\item{mask_filename}{String containing pattern pointing to zarr-lsm files}

\item{msl_pattern}{String containing pattern pointing to zarr-msl files}

\item{file_name}{File name to import}

\item{year}{year to plot}
}
\description{
R function wrapper for the asli Python package ASLICalculator functionality
See the asli Python package for full documentation
}
